rename utils file in app-pages/map/ to map-utils
