+++ src/app-pages/map/components/map-details/MapDetails.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import Accordion from "../../../../app-common/accordion/Accordion";
 import { buildLocationDetailSections } from "./data";
-import { displayTypes, defaultMapParams } from "../../utils";
+import { displayTypes, defaultMapParams } from "../../map-utils";
 import { useDimensions } from "../../../../utils";
 
 const MapDetails = ({
